==> 1_data_processing.ipynb <==
import pandas as pd

item_info_path = r'..\..\dataset\item_info.csv'
transactions_info_path = r'..\..\dataset\transactions_info.csv'


# Load item_info and transactions_info into Pandas DataFrames
item_info_df = pd.read_csv(item_info_path)
transactions_info_df = pd.read_csv(transactions_info_path)

# Merge the data based on the common item_code
merged_data = pd.merge(transactions_info_df, item_info_df, on='item_code', how='inner')

# Display the merged data (you can remove this in the final script)
print(merged_data.head())

# Save the merged data to a new CSV file
merged_data.to_csv(r'..\utils\merged_data.csv', index=False)
