==> 2_create_primary_keys.ipynb <==
import pandas as pd

# Load the merged data from the previous step
merged_data_path = r'..\utils\merged_data.csv'
merged_data = pd.read_csv(merged_data_path)

# Convert the 'invoice_time' column to a datetime object
merged_data['invoice_time'] = pd.to_datetime(merged_data['invoice_time'])

# Create new columns for primary keys (item_code, date, and hour)
merged_data['date'] = merged_data['invoice_time'].dt.date
merged_data['hour'] = merged_data['invoice_time'].dt.hour

# Display the dataset with primary keys (you can remove this in the final script)
print(merged_data.head())

# Save the dataset with primary keys to a new CSV file
merged_data.to_csv(r'DocumentsFinalsMLSalse_Forecasting_M\Salse_Forecasting_ML\src\utils\merged_data_with_keys.csv', index=False)
