==> 2_data_exploration.ipynb <==
import pandas as pd

# Define the file paths
item_info_path = r'..\dataset\item_info.csv'
transactions_info_path = r'..\dataset\transactions_info.csv'

# Load the data
item_info_df = pd.read_csv(item_info_path)
transactions_info_df = pd.read_csv(transactions_info_path)

# Data Exploration
# Display the first few rows of the loaded data frames
print("Item Information Data:")
print(item_info_df.head())

print("\nTransactions Information Data:")
print(transactions_info_df.head())

# Check data types and missing values
print("\nData Types and Missing Values in Item Information Data:")
print(item_info_df.info())

print("\nData Types and Missing Values in Transactions Information Data:")
print(transactions_info_df.info())

# Summary statistics
print("\nSummary Statistics for Item Information Data:")
print(item_info_df.describe())

print("\nSummary Statistics for Transactions Information Data:")
print(transactions_info_df.describe())
