==> 3_create_target_variable.ipynb <==
import pandas as pd

# Load the dataset with primary keys from the previous step
merged_data_path = r'..\utils\merged_data_with_keys.csv'
merged_data = pd.read_csv(merged_data_path)

# Calculate the hourly sales by shifting the 'item_qty' column
merged_data['hourly_sales'] = merged_data.groupby('item_code')['item_qty'].shift(-1)

# Display the dataset with the target variable (you can remove this in the final script)
print(merged_data.head())

# Save the dataset with the target variable to a new CSV file
merged_data.to_csv(r'..\utils\merged_data_with_target.csv', index=False)
