==> 5_feature_engineering.ipynb <==
import pandas as pd

# Define the file path
transactions_info_path = r'..\dataset\transactions_info.csv'

# Load the data
transactions_info_df = pd.read_csv(transactions_info_path)

# Feature Engineering

# Convert 'invoice_time' to a datetime object
transactions_info_df['invoice_time'] = pd.to_datetime(transactions_info_df['invoice_time'])

# Extract date-related features
transactions_info_df['invoice_date'] = transactions_info_df['invoice_time'].dt.date
transactions_info_df['day_of_week'] = transactions_info_df['invoice_time'].dt.day_name()
transactions_info_df['month'] = transactions_info_df['invoice_time'].dt.month

# Extract time-related features
transactions_info_df['hour'] = transactions_info_df['invoice_time'].dt.hour
transactions_info_df['minute'] = transactions_info_df['invoice_time'].dt.minute

# Display the updated DataFrame with new features
print("Updated Transactions Information Data:")
print(transactions_info_df.head())
