==> 5_item_related_features.ipynb <==
import pandas as pd

# Load the dataset with sales-related features from the previous steps
merged_data_path = r'..\utils\merged_data_with_sales_features.csv'
merged_data = pd.read_csv(merged_data_path)

# Load the item_info data
item_info_path = r'C:\Users\User\Salse_Forecasting_ML\dataset\item_info.csv'
item_info_df = pd.read_csv(item_info_path)

# Merge item-related features with the dataset based on the item_code
merged_data = pd.merge(merged_data, item_info_df, on='item_code', how='left')

# Display the dataset with item-related features (you can remove this in the final script)
print(merged_data.head())

# Save the dataset with item-related features to a new CSV file
merged_data.to_csv(r'..\utils\merged_data_with_item_features.csv', index=False)
