==> 6_time_related_features.ipynb <==
import pandas as pd

# Load the dataset with item-related features from the previous steps
merged_data_path = r'..\utils\merged_data_with_item_features.csv'
merged_data = pd.read_csv(merged_data_path)

# Convert the 'invoice_time' column to a datetime object
merged_data['invoice_time'] = pd.to_datetime(merged_data['invoice_time'])

# Extract time-related features
merged_data['hour_of_day'] = merged_data['invoice_time'].dt.hour
merged_data['day_of_week'] = merged_data['invoice_time'].dt.dayofweek
merged_data['month'] = merged_data['invoice_time'].dt.month

# Display the dataset with time-related features (you can remove this in the final script)
print(merged_data.head())

# Save the dataset with time-related features to a new CSV file
merged_data.to_csv(r'..\utils\merged_data_with_time_features.csv', index=False)
